Automatically re-load schema on space creation
If the space is created in one of the shards, and no requests are made
to that shard, there is no way to know that the space exists. This is
how net.box works: it only fetches schema when it sees a mismatch
between schema_id of client and server.

Then space_call() will call the callback with nil as space_obj.